feat: Prepare refactor for use of multiple entry points which send commands
