+++ src/RoadRegistry.BackOffice.Abstractions/FeatureCompare/FeatureCompareMessageRequest.cs
@@ -1,0 +1,7 @@
+namespace RoadRegistry.BackOffice.Abstractions.FeatureCompare;
+
+using RoadRegistry.BackOffice.Abstractions;
+
+public record FeatureCompareMessageRequest(string ArchiveId) : SqsMessageRequest<FeatureCompareMessageResponse>
+{
+}
