+++ src/RoadRegistry.BackOffice.Abstractions/FeatureCompare/FeatureCompareMessageResponse.cs
@@ -1,0 +1,7 @@
+namespace RoadRegistry.BackOffice.Abstractions.FeatureCompare;
+
+using RoadRegistry.BackOffice.Abstractions;
+
+public record FeatureCompareMessageResponse(ArchiveId ArchiveId) : SqsMessageResponse
+{
+}
