+++ src/RoadRegistry.BackOffice.Abstractions/SqlCommandProcessorPositionStore.cs
@@ -1,4 +1,4 @@
-namespace RoadRegistry.BackOffice.CommandHost;
+namespace RoadRegistry.BackOffice.Abstractions;
 
 using System;
 using System.Data;
