+++ src/RoadRegistry.BackOffice.Handlers/FeatureCompare/FeatureCompareMessageHandler.cs
@@ -1,13 +1,12 @@
-namespace RoadRegistry.BackOffice.MessagingHost.Sqs.Handlers
+namespace RoadRegistry.BackOffice.Handlers.FeatureCompare
 {
     using Microsoft.Extensions.Logging;
     using RoadRegistry.BackOffice.Abstractions;
     using RoadRegistry.BackOffice.Framework;
     using RoadRegistry.BackOffice.Messages;
-    using RoadRegistry.BackOffice.MessagingHost.Sqs.Requests;
-    using RoadRegistry.BackOffice.MessagingHost.Sqs.Responses;
     using System.Threading;
     using System.Threading.Tasks;
+    using Abstractions.FeatureCompare;
 
     public class FeatureCompareMessageHandler : SqsMessageRequestHandler<FeatureCompareMessageRequest, FeatureCompareMessageResponse>
     {
