--- src/RoadRegistry.BackOffice.MessagingHost.Sqs/Requests/FeatureCompareMessageRequest.cs
@@ -1,8 +1,0 @@
-namespace RoadRegistry.BackOffice.MessagingHost.Sqs.Requests;
-
-using Abstractions;
-using RoadRegistry.BackOffice.MessagingHost.Sqs.Responses;
-
-public record FeatureCompareMessageRequest(string ArchiveId) : SqsMessageRequest<FeatureCompareMessageResponse>
-{
-}
