--- src/RoadRegistry.BackOffice.MessagingHost.Sqs/Responses/FeatureCompareMessageResponse.cs
@@ -1,7 +1,0 @@
-namespace RoadRegistry.BackOffice.MessagingHost.Sqs.Responses;
-
-using Abstractions;
-
-public record FeatureCompareMessageResponse(ArchiveId ArchiveId) : SqsMessageResponse
-{
-}
